+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+#### 6.0.0
+* __Breaking change:__ Removed the legacy API for creating runtime Farkles from EGT files (the API with the transformers and fusers). EGT files are still supported (for now), but users are strongly urged to rewrite their grammars using `Farkle.Builder`, or implement the `PostProcessor` interface themselves (not recommended).
+* __Breaking change:__ The `PostProcessor` type was moved to the root `Farkle` namespace. Some reusable post-processors were moved to the new `Farkle.PostProcessors` module.
+
 #### 5.4.0 - 20-03-2020
 * You can now add lexical groups in a grammar. They resemble [GOLD Parser's feature](http://www.goldparser.org/doc/grammars/define-groups.htm), but always advance by character and do not not support nesting.
 * Add a couple of methods in runtime Farkles to easily check whether building it had succeeded.